converter/pyproject.toml
[project]
name = "functiongemma-onnx-converter"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "accelerate>=1.12.0",
    "onnx>=1.20.1",
    "onnxruntime>=1.24.2",
    "optimum[onnxruntime]>=2.1.0",
    "python-dotenv>=1.2.1",
    "torch>=2.10.0",
    "transformers>=4.57.6",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.24.2; platform_system != 'Darwin'",
]
